+++ processor/src/test/resources/bad_input/com/example/on_activate/OnActivateOnKeepAliveModel.java
@@ -1,0 +1,20 @@
+package com.example.on_activate;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Computed;
+import arez.annotations.OnActivate;
+
+@ArezComponent
+public abstract class OnActivateOnKeepAliveModel
+{
+  @Computed( keepAlive = true )
+  public int getMyValue()
+  {
+    return 0;
+  }
+
+  @OnActivate
+  void onMyValueActivate()
+  {
+  }
+}