+++ processor/src/test/resources/bad_input/com/example/on_deactivate/OnDeactivateKeepAliveModel.java
@@ -1,0 +1,20 @@
+package com.example.on_deactivate;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Computed;
+import arez.annotations.OnDeactivate;
+
+@ArezComponent
+public abstract class OnDeactivateKeepAliveModel
+{
+  @Computed( keepAlive = true )
+  public int getMyValue()
+  {
+    return 0;
+  }
+
+  @OnDeactivate
+  void onMyValueDeactivate()
+  {
+  }
+}